web_backEnd/src/ShopWeb.Tests/Service/UserServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using ShopWeb.Core.src.Entity;
using ShopWeb.Core.src.RepoAbstract;
using ShopWeb.Core.src.Role;
using ShopWeb.Service.src.DTOs;
using ShopWeb.Service.src.Services;
using Xunit;

namespace ShopWeb.Service.Tests
{
    public class UserServiceTests
    {
        private readonly UserService _userService;
        private readonly Mock<IUserRepository> _userRepositoryMock;
        private readonly Mock<IMapper> _mapperMock;

        public UserServiceTests()
        {
            _userRepositoryMock = new Mock<IUserRepository>();
            _mapperMock = new Mock<IMapper>();

            _userService = new UserService(_userRepositoryMock.Object, _mapperMock.Object);
        }

        [Fact]
        public async Task CreateUserAsync_ValidUserCreateDto_ReturnsTrue()
        {
            // Arrange
            var userCreateDto = new UserCreateDto
            {
                // Set up user create DTO
            };
            var user = new User
            {
                // Set up user
            };

            _mapperMock.Setup(x => x.Map<User>(userCreateDto)).Returns(user);
            _userRepositoryMock.Setup(x => x.CreateUserAsync(user)).ReturnsAsync(true);

            // Act
            var result = await _userService.CreateUserAsync(userCreateDto);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task DeleteUserByIdAsync_ValidUserId_ReturnsTrue()
        {
            // Arrange
            var userId = Guid.NewGuid();

            _userRepositoryMock.Setup(x => x.DeleteUserByIdAsync(userId)).ReturnsAsync(true);

            // Act
            var result = await _userService.DeleteUserByIdAsync(userId);

            // Assert
            Assert.True(result);
        }

        [Fact]
        public async Task GetAllUsersAsync_ReturnsAllUsers()
        {
            // Arrange
            var users = new List<User>
            {
                // Set up list of users
            };
            _userRepositoryMock.Setup(x => x.GetAllUsersAsync()).ReturnsAsync(users);
            _mapperMock.Setup(x => x.Map<IEnumerable<UserReadDto>>(users)).Returns(new List<UserReadDto>());

            // Act
            var result = await _userService.GetAllUsersAsync();

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetUserByIdAsync_ValidUserId_ReturnsUser()
        {
            // Arrange
            var userId = Guid.NewGuid();
            var user = new User
            {
                // Set up user
            };
            _userRepositoryMock.Setup(x => x.GetUserByIdAsync(userId)).ReturnsAsync(user);
            _mapperMock.Setup(x => x.Map<UserReadDto>(user)).Returns(new UserReadDto());

            // Act
            var result = await _userService.GetUserByIdAsync(userId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetUserByUserEmailAsync_ValidUserEmail_ReturnsUser()
        {
            // Arrange
            var userEmail = "test@example.com";
            var user = new User
            {
                // Set up user
            };
            _userRepositoryMock.Setup(x => x.GetUserByEmailAsync(userEmail)).ReturnsAsync(user);

            // Act
            var result = await _userService.GetUserByUserEmail(userEmail);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task UpdateUserByIdAsync_ValidUserUpdateDto_ReturnsTrue()
        {
            // Arrange
            var userUpdateDto = new UserUpdateDto
            {
                // Set up user update DTO
            };

            // Set up mock behavior as needed

            // Act
            var result = await _userService.UpdateUserByIdAsync(userUpdateDto);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }
    }
}
